requirements.txt
fastapi==0.115.8
httpx==0.28.1
pydantic==2.10.6
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
starlette==0.46.0
uvicorn==0.34.0
